+++ backend/services/testingService.js
@@ -1,0 +1,2 @@
+import Sandbox from 'sandbox';
+const sandbox = new Sandbox();
